+++ app/components/Transfers/DisplayUpdateBakerStake.tsx
@@ -1,12 +1,15 @@
 import React from 'react';
+import { useRouteMatch } from 'react-router';
 import { UpdateBakerStake } from '~/utils/types';
 import DisplayEstimatedFee from '~/components/DisplayEstimatedFee';
 import { useAccountName } from '~/utils/dataHooks';
-import styles from './transferDetails.module.scss';
 import { displayAsGTU } from '~/utils/gtu';
 import DisplayTransactionExpiryTime from '../DisplayTransactionExpiryTime/DisplayTransactionExpiryTime';
 import { dateFromTimeStamp } from '~/utils/timeHelpers';
 import DisplayAddress from '../DisplayAddress';
+import routes from '~/constants/routes.json';
+
+import styles from './transferDetails.module.scss';
 
 interface Props {
     transaction: UpdateBakerStake;
@@ -17,6 +20,8 @@ interface Props {
  */
 export default function DisplayUpdateBakerStake({ transaction }: Props) {
     const senderName = useAccountName(transaction.sender);
+    const isSingleSig = useRouteMatch(routes.SUBMITTRANSFER);
+
     return (
         <>
             <p className={styles.title}>From Account:</p>
@@ -30,9 +35,11 @@ export default function DisplayUpdateBakerStake({ transaction }: Props) {
                 {displayAsGTU(transaction.payload.stake)}
             </p>
             <DisplayEstimatedFee estimatedFee={transaction.estimatedFee} />
-            <DisplayTransactionExpiryTime
-                expiryTime={dateFromTimeStamp(transaction.expiry)}
-            />
+            {Boolean(isSingleSig) || (
+                <DisplayTransactionExpiryTime
+                    expiryTime={dateFromTimeStamp(transaction.expiry)}
+                />
+            )}
         </>
     );
 }