+++ app/node/nodeHelpers.ts
@@ -1,4 +1,8 @@
 import {
+    TransactionStatusEnum,
+    TransactionSummary,
+} from '@concordium/node-sdk/lib/src/types';
+import {
     getConsensusStatus,
     getAccountInfo,
     getCryptographicParameters,
@@ -6,36 +10,19 @@ import {
     getIdentityProviders,
     getAnonymityRevokers,
     getPeerList,
+    getTransactionStatus,
 } from './nodeRequests';
 import { PeerElement } from '../proto/concordium_p2p_rpc_pb';
-import { AccountInfo, Account, Global, Fraction } from '../utils/types';
-
-export interface AccountInfoPair {
-    account: Account;
-    accountInfo: AccountInfo;
-}
+import {
+    AccountInfo,
+    Global,
+    Fraction,
+    TransactionStatus,
+} from '../utils/types';
 
 export async function getlastFinalizedBlockHash(): Promise<string> {
     const consensusStatus = await getConsensusStatus();
     return consensusStatus.lastFinalizedBlock;
-}
-/** Gets the accountInfos for each given accounts. returns a list of objects
- *   each containing an account and its accountInfo.
- */
-export async function getAccountInfos(
-    accounts: Account[]
-): Promise<AccountInfoPair[]> {
-    const blockHash = await getlastFinalizedBlockHash();
-    const accountInfos: AccountInfoPair[] = await Promise.all(
-        accounts.map(async (account) => {
-            const accountInfo = await getAccountInfo(
-                account.address,
-                blockHash
-            );
-            return { account, accountInfo };
-        })
-    );
-    return accountInfos;
 }
 
 /** Gets the accountInfo for the given address. */
@@ -43,7 +30,13 @@ export async function getAccountInfoOfAddress(
     address: string
 ): Promise<AccountInfo> {
     const blockHash = await getlastFinalizedBlockHash();
-    return getAccountInfo(address, blockHash);
+    const accountInfo = await getAccountInfo(address, blockHash);
+    if (!accountInfo) {
+        throw new Error(
+            `Address ${address} does not represent an account on the connected node. Please check that your node is up to date with the blockchain.`
+        );
+    }
+    return accountInfo;
 }
 
 export async function fetchLastFinalizedBlockSummary() {
@@ -117,4 +110,47 @@ export async function isNodeUpToDate() {
 export async function nodeSupportsMemo() {
     const consensusStatus = await getConsensusStatus();
     return consensusStatus.protocolVersion >= 2;
+}
+
+export interface StatusResponse {
+    status: TransactionStatus;
+    outcomes: Record<string, TransactionSummary>;
+}
+
+/**
+ * Queries the node for the status of the transaction with the provided transaction hash.
+ * The polling will continue until the transaction becomes absent or finalized.
+ * @param transactionHash the hash of the transaction to get the status for
+ * @param pollingIntervalM, optional, interval between polling in milliSeconds, defaults to every 20 seconds.
+ */
+export async function getStatus(
+    transactionHash: string,
+    pollingIntervalMs = 20000
+): Promise<StatusResponse> {
+    return new Promise((resolve) => {
+        const interval = setInterval(async () => {
+            let response;
+            try {
+                response = await getTransactionStatus(transactionHash);
+            } catch (err) {
+                // This happens if the node cannot be reached. Just wait for the next
+                // interval and try again.
+                return;
+            }
+            // if there is no response, the transaction is absent.
+            if (!response) {
+                clearInterval(interval);
+                resolve({ status: TransactionStatus.Rejected, outcomes: {} });
+                return;
+            }
+
+            if (response.status === TransactionStatusEnum.Finalized) {
+                clearInterval(interval);
+                resolve({
+                    status: TransactionStatus.Finalized,
+                    outcomes: response.outcomes || {},
+                });
+            }
+        }, pollingIntervalMs);
+    });
 }