+++ app/pages/GenesisAccount/GenesisAccount.tsx
@@ -154,7 +154,6 @@ function ConfirmAndExport({
                 address,
                 name: accountName,
                 identityId: credentialNumberIdentityId.identityId,
-                maxTransactionId: '0',
                 isInitial: false,
                 transactionFilter: {},
             };