+++ app/pages/multisig/MultiSignatureCreateProposal/BuildProposal.tsx
@@ -79,7 +79,8 @@ export default function BuildProposal({
                             blockSummary={blockSummary}
                             consensusStatus={consensusStatus}
                         />
-                        <Form.Timestamp
+                        <Form.DatePicker
+                            className="body2 mV40"
                             name="effectiveTime"
                             label="Effective Time"
                             defaultValue={
@@ -95,8 +96,10 @@ export default function BuildProposal({
                                     'Effective time must be in the future'
                                 ),
                             }}
+                            minDate={new Date()}
                         />
-                        <Form.Timestamp
+                        <Form.DatePicker
+                            className="body2 mV40"
                             name="expiryTime"
                             label="Transaction Expiry Time"
                             defaultValue={
@@ -118,6 +121,7 @@ export default function BuildProposal({
                                     ),
                                 },
                             }}
+                            maxDate={effective ?? new Date()}
                         />
                     </>
                 </div>