+++ app/utils/rustInterface.ts
@@ -1,4 +1,5 @@
 import PromiseWorker from 'promise-worker';
+import { AccountEncryptedAmount } from '@concordium/node-sdk/';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-expect-error : has no default export.
 import RustWorker, { BakerKeyVariants } from './rust.worker';
@@ -11,7 +12,6 @@ import {
     CredentialDeploymentDetails,
     CredentialDeploymentInformation,
     Global,
-    AccountEncryptedAmount,
     GenesisAccount,
     SignedIdRequest,
     UnsignedCredentialDeploymentInformation,
@@ -350,9 +350,10 @@ export async function makeTransferToPublicData(
         accountNumber,
         incomingAmounts: accountEncryptedAmount.incomingAmounts,
         encryptedSelfAmount: accountEncryptedAmount.selfAmount,
-        aggIndex:
+        aggIndex: (
             accountEncryptedAmount.startIndex +
-            accountEncryptedAmount.incomingAmounts.length,
+            BigInt(accountEncryptedAmount.incomingAmounts.length)
+        ).toString(),
     };
 
     const transferToPublicData = await worker.postMessage({
@@ -376,9 +377,6 @@ export async function makeTransferToEncryptedData(
         accountNumber,
         incomingAmounts: accountEncryptedAmount.incomingAmounts,
         encryptedSelfAmount: accountEncryptedAmount.selfAmount,
-        aggIndex:
-            accountEncryptedAmount.startIndex +
-            accountEncryptedAmount.incomingAmounts.length,
     };
 
     const transferToSecretData = await worker.postMessage({
@@ -404,9 +402,10 @@ export async function makeEncryptedTransferData(
         accountNumber,
         incomingAmounts: accountEncryptedAmount.incomingAmounts,
         encryptedSelfAmount: accountEncryptedAmount.selfAmount,
-        aggIndex:
+        aggIndex: (
             accountEncryptedAmount.startIndex +
-            accountEncryptedAmount.incomingAmounts.length,
+            BigInt(accountEncryptedAmount.incomingAmounts.length)
+        ).toString(),
     };
 
     const encryptedTransferData = await worker.postMessage({
@@ -499,7 +498,7 @@ export function getAddressFromCredentialId(credId: string): Promise<string> {
     });
 }
 
-export function getCredId(
+export function computeCredId(
     prfKeySeed: string,
     credentialNumber: number,
     global: Global