+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## Unreleased
 
+### Added
+
+-   Added a GTU drop option for testnet and stagenet.
+-   In the case of a failed identity, the error details received from the identity provider are now displayed to the user.
+
 ### Changed
 
+-   Auxiliary data in an Update Protocol transaction is now optional.
+-   Updated terms and conditions.
 -   Updated UI to reflect the rename of GTU to CCD, meaning anywhere tokens were referred to as GTU, it now says CCD. The GTU icon has also been replaced with the icon representing CCD.
 
 ## 1.2.0
@@ -24,6 +31,7 @@
 -   It is now possible to set a default account, which will be selected initially when opening the app.
 -   "About" menu link has been added to native menu bar under "Help" for windows/linux. It is already accessible on mac under "Concordium desktop wallet" in the native menu.
 -   When an account is synchronizing transactions, it now shows with a spinning icon in the tab header of the transaction log.
+-   Added flows to make baker transactions from accounts with only one set of credentials possible without going through a multi-signature proposal flow.
 
 ### Changed